refactor: rename DataStore to AppPreferences and simplify methods
